EXCEPTION_INVALID_HANDLE has its high bit set, so cast ex to unsigned.
Fixes

  error(clang): case value evaluates to 3221225480, which cannot be narrowed to type 'int' [-Wc++11-narrowing]

BUG=chromium:82385
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/349233005

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13405 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: 1e0e8224bfc9d3ed0d7b35b670fa6ab6f19e7d19

+++ src/trusted/debug_stub/win/thread_impl.cc
@@ -28,7 +28,7 @@ enum PosixSignals {
 };
 
 int IThread::ExceptionToSignal(int ex) {
-  switch (ex) {
+  switch (static_cast<DWORD>(ex)) {
     case EXCEPTION_GUARD_PAGE:
     case EXCEPTION_ARRAY_BOUNDS_EXCEEDED:
     case EXCEPTION_DATATYPE_MISALIGNMENT:
